# docs/swagger.yaml
definitions:
  asset_controllers.AssetsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.AsLand'
        type: array
      message:
        example: List all assets
        type: string
    type: object
  asset_controllers.ErrorResponse:
    properties:
      error:
        example: Detailed error description
        type: string
      message:
        example: Error message
        type: string
    type: object
  asset_controllers.GetAssetResponse:
    properties:
      data:
        $ref: '#/definitions/models.AsLand'
      message:
        example: Asset retrieved successfully
        type: string
    type: object
  models.AsLand:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      cutax_land_id:
        type: string
      deed_no:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      estimated_price:
        type: number
      estimated_price_by_treasury:
        type: number
      id:
        type: string
      is_estimated_by_treasury:
        type: boolean
      land_district_id:
        type: integer
      land_no:
        type: string
      land_space_ngan:
        type: number
      land_space_rai:
        type: number
      land_space_sub_wa:
        type: number
      land_space_wa:
        type: number
      land_sub_district_id:
        type: integer
      land_zone:
        type: string
      map_geometry:
        type: string
      map_lat:
        type: string
      map_long:
        type: string
      muni_code:
        type: string
      note:
        type: string
      parcel_no:
        type: string
      parcel_type:
        type: integer
      road:
        type: string
      source_dt:
        type: string
      special_usetax_type:
        type: string
      survey_no:
        type: string
      sync_id:
        type: string
      tax_year:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      utm_map1:
        type: string
      utm_map2:
        type: string
      utm_map3:
        type: string
      utm_map4:
        type: string
      utm_scale:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/assets:
    get:
      consumes:
      - application/json
      description: Get a list of all assets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset_controllers.AssetsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/asset_controllers.ErrorResponse'
      summary: List all assets
      tags:
      - assets
  /api/v1/assets/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an asset by its ID
      parameters:
      - description: Asset ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/asset_controllers.GetAssetResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/asset_controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/asset_controllers.ErrorResponse'
      summary: Get asset by ID
      tags:
      - assets
swagger: "2.0"